Extract DB factory and add migration tests
